+++ petitparser/lib/src/parser/repeater/repeating.dart
@@ -4,11 +4,8 @@ import 'unbounded.dart';
 
 /// An abstract parser that repeatedly parses between 'min' and 'max' instances
 /// of its delegate.
-abstract class RepeatingParser<T> extends DelegateParser<T, List<T>> {
-  final int min;
-  final int max;
-
-  RepeatingParser(Parser<T> parser, this.min, this.max) : super(parser) {
+abstract class RepeatingParser<R> extends DelegateParser<R, List<R>> {
+  RepeatingParser(Parser<R> parser, this.min, this.max) : super(parser) {
     if (min < 0) {
       throw ArgumentError(
           'Minimum repetitions must be positive, but got $min.');
@@ -19,11 +16,17 @@ abstract class RepeatingParser<T> extends DelegateParser<T, List<T>> {
     }
   }
 
+  /// The minimum amount of repetitions.
+  final int min;
+
+  /// The maximum amount of repetitions, or [unbounded].
+  final int max;
+
   @override
   String toString() =>
       '${super.toString()}[$min..${max == unbounded ? '*' : max}]';
 
   @override
-  bool hasEqualProperties(RepeatingParser<T> other) =>
+  bool hasEqualProperties(RepeatingParser<R> other) =>
       super.hasEqualProperties(other) && min == other.min && max == other.max;
 }